examples: reorganized examples' sweepers and utils into new contrib dir
this enables reusablity of those sweepers without copy&paste and without
inter-referencing different examples
