+++ examples/vanderpol/vdp_sdc.cpp
@@ -9,7 +9,7 @@
 #include <pfasst/controller/sdc.hpp>
 #include <pfasst/encap/vector.hpp>
 
-#include "vdp_sweeper.hpp"
+#include <pfasst/contrib/sweeper/vdp_sweeper.hpp>
 
 namespace pfasst
 {
